# WARNING: head commit changed in the meantime
item added

+++ src/main/java/PlayerCharacter.java
@@ -2,7 +2,7 @@
 public class PlayerCharacter extends Creature {
 	private int strength, stamina, agility, dodgeRating;
 
-	public PlayerCharacter(String characterName, int life, double speed, int level) {
+	public PlayerCharacter(String characterName, int life, int speed, int level) {
 		super(characterName, life, speed, level);
 		this.strength = 1;
 		this.stamina = 1;
